+++ paper-server/patches/sources/net/minecraft/world/entity/ai/behavior/TransportItemsBetweenContainers.java.patch
@@ -1,27 +1,14 @@
 --- a/net/minecraft/world/entity/ai/behavior/TransportItemsBetweenContainers.java
 +++ b/net/minecraft/world/entity/ai/behavior/TransportItemsBetweenContainers.java
-@@ -323,11 +_,23 @@
+@@ -323,7 +_,10 @@
              } else {
                  boolean flag1 = this.isWantedBlock(mob, transportItemTarget.state)
                      && !this.isPositionAlreadyVisited(visited, unreachable, transportItemTarget, level)
 -                    && !this.isContainerLocked(transportItemTarget);
-+                    // Paper start - CopperGolemValidateTargetEvent
++                    // Paper start - ItemTransportingEntityValidateTargetEvent
 +                    && !this.isContainerLocked(transportItemTarget)
-+                    && this.callValidateEvent(mob, level, transportItemTarget);
-+                    // Paper end - CopperGolemValidateTargetEvent
++                    && org.bukkit.craftbukkit.event.CraftEventFactory.callTransporterValidateTarget(mob, level, transportItemTarget.pos);
++                    // Paper end - ItemTransportingEntityValidateTargetEvent
                  return flag1 ? transportItemTarget : null;
              }
          }
-     }
-+
-+    // Paper start - CopperGolemValidateTargetEvent
-+    private boolean callValidateEvent(final PathfinderMob mob, final Level level, final TransportItemsBetweenContainers.TransportItemTarget transportItemTarget) {
-+        if (mob.getType() != net.minecraft.world.entity.EntityType.COPPER_GOLEM) return true; // In case a plugin added this task to another mob
-+        final var event = new io.papermc.paper.event.entity.CopperGolemValidateTargetEvent((org.bukkit.entity.CopperGolem) mob.getBukkitEntity(), new org.bukkit.craftbukkit.block.CraftBlock(level, transportItemTarget.pos));
-+        event.callEvent();
-+        return event.isAllowed();
-+    }
-+    // Paper end - CopperGolemValidateTargetEvent
- 
-     private boolean isContainerLocked(TransportItemsBetweenContainers.TransportItemTarget target) {
-         return target.blockEntity instanceof BaseContainerBlockEntity baseContainerBlockEntity && baseContainerBlockEntity.isLocked();
